Backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
Pillow
python-multipart
httpx
numpy
torch
torchvision